DjangoRestFramework/requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==7.1.2
click-aliases==1.0.1
click-default-group==1.2.4
click-help-colors==0.9.2
colorama==0.4.6
commonmark==0.9.1
Django==4.2.4
django-compat==1.0.15
django-cors-headers==4.2.0
django-graphql-jwt==0.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
graphene==3.3
graphene-django==3.1.5
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==2.0.2
icalendar==5.0.9
idna==3.4
ply==3.11
promise==2.3
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
rich==9.13.0
schedule==1.2.0
six==1.16.0
SQLAlchemy==1.4.49
sqlparse==0.4.4
tabulate==0.8.10
task==0.2.5
text-unidecode==1.3
thriftpy==0.3.9
total==0.0.7
typing-extensions==3.10.0.2
tzdata==2023.3
urllib3==2.0.5
